Conflicts:
	src/main/java/com/workingflows/js/jscore/client/api/db/IDBEnvironment.java
	src/main/java/com/workingflows/js/jscore/client/api/db/IDBFactory.java
	src/main/java/com/workingflows/js/jscore/client/api/db/IDBOpenDBRequest.java

+++ src/main/java/com/workingflows/js/jscore/client/api/db/IDBFactory.java
@@ -14,7 +14,7 @@ import com.google.gwt.core.client.js.JsType;
 @JsType(prototype = "IDBFactory")
 public interface IDBFactory {
 
-    public IDBOpenDBRequest open(String name, long version);
+    public IDBOpenDBRequest open(String name, int version);
 
     public IDBOpenDBRequest deleteDatabase(String name);
